Changed the intermediate buffer from short ints to ints

+++ single_header/simpleALSA.h
@@ -112,7 +112,7 @@ struct sa_device
     snd_pcm_sw_params_t *swparams;
 
     /** Pointer to the place is memory where audio samples are written right before being send to the ALSA buffer */
-    signed short *samples;
+    int *samples;
 
     /** Indicates support for the hardware to pause the pcm stream */
     bool supportsPause;
@@ -533,7 +533,7 @@ sa_result init_alsa_device(sa_device *device) {
         exit(EXIT_FAILURE);
     }
 
-    device->samples = (signed short *) malloc((device->periodSize * device->config->channels *
+    device->samples = (int *) malloc((device->periodSize * device->config->channels *
                                                snd_pcm_format_physical_width(device->config->format)) /
                                               8);
 
@@ -852,7 +852,7 @@ sa_result close_playback_thread(sa_device *device) {
 }
 
 sa_result write_and_poll_loop(sa_device *device, sa_poll_management *poll_manager) {
-    signed short *ptr;
+    int *ptr;
     int err, cptr, init, readcount;
     readcount = 1;
     init      = 1;
